feat: enhance api key revocation flow with owner lookup and security notification

+++ src/main/java/com/guts/Guts_IAM/apikey/model/ApiKey.java
@@ -28,7 +28,7 @@ public class ApiKey {
     private String hashedApiKey;
 
 
-    private long userId;
+    private Integer userId;
 
     private Status keyStatus=Status.ACTIVE;
